Unified both regression test helper files using conditional compilation #3504
